db_setup/requirements.txt
SQLAlchemy
psycopg2-binary
pgvector
openai
python-dotenv
requests
icalendar